+++ oml/__init__.py
@@ -1,3 +1,3 @@
 "Object moderation layer for Django models"
 
-__version__ = '0.1.0'
+__version__ = '0.1.1'